Refactor: Separate report and service modules
There's a distinction between reports and services. Reports post data
back to Scholarsphere, whereas the service generates the data that goes
into the report. Typically, there's one report per service, but there
doesn't have to be. Refactoring clarifies which kind of class does
which.
