+++ lib/metadata_listener.rb
@@ -1,14 +1,14 @@
 # frozen_string_literal: true
 
 module MetadataListener
-  require 'metadata_listener/redis'
   require 'metadata_listener/job'
+  require 'metadata_listener/redis'
+  require 'metadata_listener/report/metadata'
+  require 'metadata_listener/report/virus'
   require 'metadata_listener/s3_client'
-  require 'metadata_listener/fits_service'
-  require 'metadata_listener/tika'
-  require 'metadata_listener/clamav_service'
-  require 'metadata_listener/virus_reporting_service'
-  require 'metadata_listener/metadata_reporting_service'
+  require 'metadata_listener/service/clamav'
+  require 'metadata_listener/service/fits'
+  require 'metadata_listener/service/tika'
 
   class << self
     # @note Because this is intended to run inside a Docker container, it is customary to write logs to STDOUT
