+++ lib/metadata_listener/job.rb
@@ -9,13 +9,13 @@ module MetadataListener
     # @param [String] path indicating where the file is stored in S3
     # @param [String] endpoint to send results to
     # @param [String] api_token used to authenticate with endpoint
-    # @param [Array<Symbol>] services to run, defaults to []
+    # @param [Array<Symbol>] reporting services to run, defaults to []
     def perform(path:, endpoint: nil, api_token: nil, services: [])
       file = MetadataListener.s3_client.download_file(path)
       services << :virus if services.empty?
 
       services.map do |key|
-        available_services[key].call(path: file.body.path, endpoint: endpoint, api_token: api_token)
+        available_reports[key].call(path: file.body.path, endpoint: endpoint, api_token: api_token)
       end
 
       FileUtils.rm_f(file.body.path)
@@ -23,10 +23,10 @@ module MetadataListener
 
     private
 
-      def available_services
+      def available_reports
         {
-          virus: VirusReportingService,
-          metadata: MetadataReportingService
+          virus: Report::Virus,
+          metadata: Report::Metadata
         }
       end
   end
