+++ spec/fixtures/vcr_cassettes/MetadataListener_Report_Metadata/when_reporting_results/sends_the_results_to_an_endpoint.yml
