+++ spec/fixtures/vcr_cassettes/MetadataListener_Report_Metadata/when_unable_to_report_results/logs_the_response_from_the_endpoint.yml
