+++ spec/lib/metadata_listener/job_spec.rb
@@ -4,17 +4,17 @@ RSpec.describe MetadataListener::Job do
   let(:path) { upload_file(file: fixture_path.join('1.pdf')) }
 
   before do
-    allow(MetadataListener::VirusReportingService).to receive(:call)
-    allow(MetadataListener::MetadataReportingService).to receive(:call)
+    allow(MetadataListener::Report::Virus).to receive(:call)
+    allow(MetadataListener::Report::Metadata).to receive(:call)
   end
 
   it { is_expected.to be_an(ActiveJob::Base) }
 
   context 'without requesting any specific services' do
-    it 'calls the VirusReportingService by default' do
+    it 'calls the virus report by default' do
       described_class.perform_now(path: path)
-      expect(MetadataListener::MetadataReportingService).not_to have_received(:call)
-      expect(MetadataListener::VirusReportingService).to have_received(:call).with(
+      expect(MetadataListener::Report::Metadata).not_to have_received(:call)
+      expect(MetadataListener::Report::Virus).to have_received(:call).with(
         path: a_kind_of(String),
         endpoint: nil,
         api_token: nil
@@ -23,10 +23,10 @@ RSpec.describe MetadataListener::Job do
   end
 
   context 'when requesting a metadata report' do
-    it 'calls the Metadata report' do
+    it 'calls the metadata report' do
       described_class.perform_now(path: path, services: [:metadata])
-      expect(MetadataListener::VirusReportingService).not_to have_received(:call)
-      expect(MetadataListener::MetadataReportingService).to have_received(:call).with(
+      expect(MetadataListener::Report::Virus).not_to have_received(:call)
+      expect(MetadataListener::Report::Metadata).to have_received(:call).with(
         path: a_kind_of(String),
         endpoint: nil,
         api_token: nil
